Add files via upload
In this tutorial, we will learn to make a Quote website using HTML CSS, and JavaScript. Every time people will visit this website it will display a new Quote. In this code generator website, we will add a button that will generate new quotes when clicked on it. and there will be one more button to share this quote on Twitter. We will make this Quote Generator website using HTML and JavaScript